backend/pyproject.toml
[project]
name = "web-llm-template-backend"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",

    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "httpx>=0.27.0",

    # Config
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",

    # Redis
    "redis>=5.0.0",

    # Google Cloud
    "google-cloud-storage>=2.16.0",
    "google-cloud-firestore>=2.16.0",
    "google-genai>=1.0.0",

    # Email (Brevo)
    "brevo-python>=1.0.0",

    # Utilities
    "python-slugify>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
